pyproject.toml
[tool.poetry]
name = "qualitative-analysis"
version = "0.1.0"
description = "qualitative analysis for master thesis"
authors = ["opmvpc <thibsix@outlook.be>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
nltk = "^3.8.1"
python-docx = "^1.1.2"
plotext = "^5.2.8"
rich = "^13.7.1"
textblob = "^0.18.0.post0"
scikit-learn = "^1.5.1"
gensim = "^4.3.3"
networkx = "^3.3"
transformers = "^4.43.4"
torch = "^2.4.0"
sentencepiece = "^0.2.0"
openai = "^1.39.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
